requirements.txt
argparse
biopython
click
easy-parallel
h5py
numpy
pandas
python-dotenv
scipy
sklearn
tables
torch
tqdm
